renamed cnmf.py to model.py
